+++ src/js/media/media.js
@@ -510,7 +510,14 @@ vjs.MediaTechController.withSourceHandlers = function(Tech){
     this.disposeSourceHandler();
     this.off('dispose', this.disposeSourceHandler);
 
-    this.currentSource_ = source;
+    // Set currentSource_ asynchronously to simulate the media element's
+    // asynchronous execution of the `resource selection algorithm`
+    this.setTimeout(vjs.bind(this, function () {
+      if (source && source.src !== '') {
+        this.currentSource_ = source;
+      }
+    }), 0);
+
     this.sourceHandler_ = sh.handleSource(source, this);
     this.on('dispose', this.disposeSourceHandler);